updated
added graphics n stuff. lots of time spent. Idk. Its late. I need to come back to this later. current problem: if I build up the 'players' dataframe, and use that, it works fine. If I save said df as a csv and then read it in, it doesnt work. Why? Wtf. this is stupid.